requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
autopep8==1.7.0
bcrypt==4.2.0
beautifulsoup4==4.12.2
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.2.3
cssselect==1.2.0
distlib==0.3.6
dnspython==2.3.0
docker==6.0.1
duckduckgo-search==2.8.6
email_validator==2.2.0
exceptiongroup==1.1.1
executing==2.1.0
fastapi==0.115.0
fastapi-cli==0.0.5
filelock==3.11.0
flake8==6.0.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
greenlet==3.1.0
gTTS==2.3.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
icecream==2.1.3
identify==2.5.22
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.4
loguru==0.7.0
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
orjson==3.8.10
outcome==1.2.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
pendulum==3.0.0
Pillow==9.5.0
pinecone-client==2.2.1
pio==0.0.3
platformdirs==3.2.0
playsound==1.2.2
pluggy==1.0.0
polars==1.7.1
pre-commit==3.2.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.0.1
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.3.1
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.9
PyYAML==6.0
readability-lxml==0.8.1
redis==4.5.4
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.8.1
rsa==4.9
selenium==4.8.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sourcery==1.2.0
SQLAlchemy==2.0.35
sqlmodel==0.0.22
starlette==0.38.5
tiktoken==0.3.3
time-machine==2.15.0
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.13.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.30.6
#uvloop==0.20.0
virtualenv==20.21.0
watchfiles==0.24.0
webdriver-manager==3.8.6
websocket-client==1.5.1
websockets==13.0.1
win32-setctime==1.1.0
wsproto==1.2.0
yarl==1.8.2
fastapi-pagination==0.12.29